refactor 패키지 관심사별 분리
