+++ src/test/java/nextstep/subway/applicaion/LineServiceTest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.unit;
+package nextstep.subway.applicaion;
 
 import nextstep.subway.applicaion.LineService;
 import nextstep.subway.applicaion.dto.LineRequest;
